# [SPARK-49990][SQL] Improve performance of `randStr`