--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/randomExpressions.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/randomExpressions.scala
@@ -28,7 +28,6 @@ import org.apache.spark.sql.catalyst.trees.{BinaryLike, TernaryLike, UnaryLike}
 import org.apache.spark.sql.catalyst.trees.TreePattern.{EXPRESSION_WITH_RANDOM_SEED, RUNTIME_REPLACEABLE, TreePattern}
 import org.apache.spark.sql.errors.QueryCompilationErrors
 import org.apache.spark.sql.types._
-import org.apache.spark.unsafe.types.UTF8String
 import org.apache.spark.util.random.XORShiftRandom
 
 /**
@@ -403,22 +402,7 @@ case class RandStr(
 
   override def evalInternal(input: InternalRow): Any = {
     val numChars = length.eval(input).asInstanceOf[Number].intValue()
-    val bytes = new Array[Byte](numChars)
-    (0 until numChars).foreach { i =>
-      // We generate a random number between 0 and 61, inclusive. Between the 62 different choices
-      // we choose 0-9, a-z, or A-Z, where each category comprises 10 choices, 26 choices, or 26
-      // choices, respectively (10 + 26 + 26 = 62).
-      val num = (rng.nextInt() % 62).abs
-      num match {
-        case _ if num < 10 =>
-          bytes.update(i, ('0' + num).toByte)
-        case _ if num < 36 =>
-          bytes.update(i, ('a' + num - 10).toByte)
-        case _ =>
-          bytes.update(i, ('A' + num - 36).toByte)
-      }
-    }
-    UTF8String.fromBytes(bytes)
+    ExpressionImplUtils.randStr(rng, numChars)
   }
 
   override def doGenCode(ctx: CodegenContext, ev: ExprCode): ExprCode = {
@@ -426,25 +410,12 @@ case class RandStr(
     val rngTerm = ctx.addMutableState(className, "rng")
     ctx.addPartitionInitializationStatement(
       s"$rngTerm = new $className(${seed}L + partitionIndex);")
-    val bytesTerm = ctx.freshName("bytes")
-    val i = ctx.freshName("i")
-    val v = ctx.freshName("v")
     val eval = length.genCode(ctx)
     ev.copy(code =
       code"""
         |${eval.code}
-        |byte[] $bytesTerm = new byte[(int)(${eval.value})];
-        |for (int $i = 0; $i < $bytesTerm.length; $i++) {
-        |  int $v = Math.abs($rngTerm.nextInt() % 62);
-        |  if ($v < 10) {
-        |    $bytesTerm[$i] = (byte)('0' + $v);
-        |  } else if ($v < 36) {
-        |    $bytesTerm[$i] = (byte)('a' + ($v - 10));
-        |  } else {
-        |    $bytesTerm[$i] = (byte)('A' + ($v - 36));
-        |  }
-        |}
-        |UTF8String ${ev.value} = UTF8String.fromBytes($bytesTerm);
+        |UTF8String ${ev.value} =
+        |  ${classOf[ExpressionImplUtils].getName}.randStr($rngTerm, (int)(${eval.value}));
         |boolean ${ev.isNull} = false;
         |""".stripMargin,
       isNull = FalseLiteral)