### What changes were proposed in this pull request?
The pr aims to 
- improve `performance` of `randStr`.
- make the `non-codegen` and `codegen` implementations of `randStr` as consistent as possible.

### Why are the changes needed?
Obviously, when creating `UTF8String`, using `UTF8String.fromBytes(bytes)` performs better than using `UTF8String.fromString(new String(bytes))`.
   
- `UTF8String.fromBytes(bytes)`
https://github.com/apache/spark/blob/39112e4f2f8c1401ffa73c84398d3b8f0afa211a/common/unsafe/src/main/java/org/apache/spark/unsafe/types/UTF8String.java#L160-L170

- `UTF8String.fromString(new String(bytes))`
https://github.com/apache/spark/blob/39112e4f2f8c1401ffa73c84398d3b8f0afa211a/common/unsafe/src/main/java/org/apache/spark/unsafe/types/UTF8String.java#L179-L184

### Does this PR introduce _any_ user-facing change?
No.


### How was this patch tested?
Manually check.


### Was this patch authored or co-authored using generative AI tooling?
No.