@cloud-fan 2024-10-18
Can we rewrite `RandStr` to a new expression that extends `Nondeterministic`? We can use `InheritAnalysisRules` so that the `Nondeterministic` expression is still available in the expression tree.