@cloud-fan 2024-10-18
actually, nvm. The issue is we must implement codegen manually for `Nondeterministic`. It won't make much difference if we use a smaller expression to extend `Nondeterministic`